Convert Nones to NaNs via .astype(float)
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CPAnalyst@8721 1fc53939-2000-0410-845c-e8453a809027

+++ src/Scatter.py
@@ -187,7 +187,7 @@ class ScatterPanel(PlotPanel):
             self.subplot = self.figure.add_subplot(111)
         self.subplot.clear()
         for i, pt_list in enumerate(self.point_lists):
-            plot_pts = np.array(pt_list)
+            plot_pts = np.array(pt_list).astype(float)
             
             if len(plot_pts)==0:
                 logging.error('No points to plot!')
